+++ scripts/west_commands/export_app/Changelog.md
@@ -1,6 +1,16 @@
 # Export App
 
+## [25.09.00]
+
+- New Features
+  - Support new option "--bf", this will copy board files to the output directory.
+  - Support set output_dir and clean_output_dir in west config file.
+
+- Improvements
+  - Make build command an instance function of CMakeApp.
+
 ## [25.06.00]
+
 - New Features
   - Support export common app without board/core variable.
   - Support dsp examples which include another sysbuild.cmake in example root's one.
@@ -16,4 +26,5 @@
   - Fixed failured builds case by case.
 
 ## [25.03.00]
+
 - The initial version.